requirements.txt
streamlit
stravalib
scikit-learn
pandas
folium
streamlit-folium
plotly
python-dotenv
geopy
polyline
